Split auth.py into several files
