+++ src/acanban/auth/__init__.py
@@ -1,0 +1,5 @@
+from .user import User  # noqa
+from .authenticate import Authenticator
+from .routes import blueprint
+
+__all__ = ['Authenticator', 'blueprint']
